Revert ID type changes

+++ internal/model/base.go
@@ -6,7 +6,7 @@ import (
 
 // BaseModel contains common fields for all models
 type BaseModel struct {
-	ID        int64     `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID        int       `gorm:"primaryKey;autoIncrement" json:"id"`
 	CreatedAt time.Time `gorm:"autoCreateTime" json:"created_at"`
 	UpdatedAt time.Time `gorm:"autoUpdateTime" json:"updated_at"`
 }
